Splitting HomeScreenCard Part 1
