+++ lib/constants/country_codes.dart
@@ -1,0 +1,10 @@
+import 'package:linum/frontend_functions/country_flag_generator.dart';
+
+final Map<String, String> countryFlagsToCountryCode =
+{
+  countryFlag('de'): "de",
+  countryFlag('gb'): "en",
+  countryFlag('nl'): "nl",
+  countryFlag('es'): "es",
+  countryFlag('fr'): "fr",
+};
