+++ lib/models/home_screen_card_data.dart
@@ -1,0 +1,11 @@
+class HomeScreenCardData {
+  final num balance;
+  final num income;
+  final num expense;
+
+  const HomeScreenCardData({
+  required this.balance,
+  required this.income,
+  required this.expense,
+  });
+}
